Rename concurrent_queue_internal to concurrent_queue
