+++ cpp/src/arrow/acero/concurrent_queue.h
